Move RolesDefinitions to testFixtures
Preparation to add more stuff to have cleaner tests, as `Roles.roles()`
will return a `Map<String, Role>` instead of `List<Role>`.
