+++ server/src/test/java/io/crate/role/TransportRoleActionTest.java
@@ -21,12 +21,12 @@
 
 package io.crate.role;
 
+import static io.crate.role.RolesDefinitions.DUMMY_USERS;
+import static io.crate.role.RolesDefinitions.DUMMY_USERS_AND_ROLES;
+import static io.crate.role.RolesDefinitions.SINGLE_USER_ONLY;
+import static io.crate.role.RolesDefinitions.getSecureHash;
+import static io.crate.role.RolesDefinitions.usersMetadataOf;
 import static io.crate.testing.Asserts.assertThat;
-import static io.crate.role.metadata.RolesDefinitions.DUMMY_USERS;
-import static io.crate.role.metadata.RolesDefinitions.DUMMY_USERS_AND_ROLES;
-import static io.crate.role.metadata.RolesDefinitions.SINGLE_USER_ONLY;
-import static io.crate.role.metadata.RolesDefinitions.getSecureHash;
-import static io.crate.role.metadata.RolesDefinitions.usersMetadataOf;
 
 import java.util.Map;
