+++ server/src/test/java/io/crate/role/metadata/UsersMetadataTest.java
@@ -28,7 +28,6 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 
-
 import org.elasticsearch.common.Strings;
 import org.elasticsearch.common.io.stream.BytesStreamOutput;
 import org.elasticsearch.common.io.stream.StreamInput;
@@ -41,6 +40,7 @@ import org.elasticsearch.test.ESTestCase;
 import org.junit.Test;
 
 import io.crate.role.Role;
+import io.crate.role.RolesDefinitions;
 import io.crate.role.SecureHash;
 
 public class UsersMetadataTest extends ESTestCase {
