+++ server/src/testFixtures/java/io/crate/role/RolesDefinitions.java
@@ -6,7 +6,7 @@
  * you may not use this file except in compliance with the License.  You may
  * obtain a copy of the License at
  *
- *   http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
@@ -19,7 +19,8 @@
  * software solely pursuant to the terms of the relevant commercial agreement.
  */
 
-package io.crate.role.metadata;
+package io.crate.role;
+
 
 import static io.crate.testing.Asserts.assertThat;
 
@@ -30,8 +31,7 @@ import java.util.Map;
 
 import org.elasticsearch.common.settings.SecureString;
 
-import io.crate.role.Role;
-import io.crate.role.SecureHash;
+import io.crate.role.metadata.UsersMetadata;
 
 public final class RolesDefinitions {
